Create First and Last Touch Queries
The code demonstrates my proficiency in SQL fundamentals, including selecting and counting distinct values, using subqueries, and performing aggregations and filtering. I analyze a "page_visits" table, determining the number of distinct campaigns and sources, the pages visited, and calculating first and last touch metrics for each campaign. The analysis reveals that the "weekly-newsletter" and "retargetting-ad" campaigns were the most successful in driving purchases, while the "interview-with-cool-tshirts-founder" and "cool-tshirts-search" campaigns underperformed. Based on these insights, I provide a recommendation to continue the top-performing campaigns and discontinue the lower-performing ones. This is a Codecademy project. All code is my own.